+++ tests/add-fscomp2.sh
@@ -37,16 +37,42 @@ test_pot_add_fscomp_101()
 	assertEquals "echo arg" "$POT_FS_ROOT/fscomp/test-fscomp $POT_FS_ROOT/jails/test-pot/m/test-mnt" "$ECHO_CALL1_ARG1"
 
 	setUp
+	_add_f_to_p test-fscomp test-pot test-mnt NO
+	assertEquals "echo calls" "1" "$ECHO_CALLS"
+	assertEquals "zfs calls" "0" "$ZFS_CALLS"
+	assertEquals "echo arg" "$POT_FS_ROOT/fscomp/test-fscomp $POT_FS_ROOT/jails/test-pot/m/test-mnt" "$ECHO_CALL1_ARG1"
+
+	setUp
 	_add_f_to_p /zroot/test-fscomp test-pot test-mnt external
 	assertEquals "echo calls" "1" "$ECHO_CALLS"
 	assertEquals "zfs calls" "0" "$ZFS_CALLS"
 	assertEquals "echo arg" "/zroot/test-fscomp $POT_FS_ROOT/jails/test-pot/m/test-mnt" "$ECHO_CALL1_ARG1"
+}
 
-	setUp
-	_add_f_to_p zroot/test-fscomp test-pot test-mnt external
+test_pot_add_fscomp_102()
+{
+	_add_f_to_p test-fscomp test-pot test-mnt NO ro
 	assertEquals "echo calls" "1" "$ECHO_CALLS"
 	assertEquals "zfs calls" "0" "$ZFS_CALLS"
-	assertEquals "echo arg" "/zroot/test-fscomp $POT_FS_ROOT/jails/test-pot/m/test-mnt" "$ECHO_CALL1_ARG1"
+	assertEquals "echo arg" "$POT_FS_ROOT/fscomp/test-fscomp $POT_FS_ROOT/jails/test-pot/m/test-mnt ro" "$ECHO_CALL1_ARG1"
+
+	setUp
+	_add_f_to_p test-fscomp test-pot test-mnt NO zfs-remount
+	assertEquals "echo calls" "1" "$ECHO_CALLS"
+	assertEquals "zfs calls" "0" "$ZFS_CALLS"
+	assertEquals "echo arg" "$POT_FS_ROOT/fscomp/test-fscomp $POT_FS_ROOT/jails/test-pot/m/test-mnt zfs-remount" "$ECHO_CALL1_ARG1"
+
+	setUp
+	_add_f_to_p /zroot/test-fscomp test-pot test-mnt external ro
+	assertEquals "echo calls" "1" "$ECHO_CALLS"
+	assertEquals "zfs calls" "0" "$ZFS_CALLS"
+	assertEquals "echo arg" "/zroot/test-fscomp $POT_FS_ROOT/jails/test-pot/m/test-mnt ro" "$ECHO_CALL1_ARG1"
+
+	setUp
+	_add_f_to_p /zroot/test-fscomp test-pot test-mnt external zfs-remount
+	assertEquals "echo calls" "1" "$ECHO_CALLS"
+	assertEquals "zfs calls" "0" "$ZFS_CALLS"
+	assertEquals "echo arg" "/zroot/test-fscomp $POT_FS_ROOT/jails/test-pot/m/test-mnt zfs-remount" "$ECHO_CALL1_ARG1"
 }
 
 setUp()