+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- add-fscomp: add the ability to remount a fscomp (-w), instead of mount it via nullfs
+- add-fscomp: add the ability to mount a fscomp in read-only (-r)
+
 ### Changed
 - private datasets are now mounted changing the mountpoint of the dataset, instead of relying on nullfs
   For that reasons, an option called "zfs-remount" is added to the fs.conf file